+++ leather/chart.py
@@ -2,6 +2,7 @@
 
 from copy import copy
 import os
+import warnings
 import xml.etree.ElementTree as ET
 
 import six
@@ -13,7 +14,7 @@ from leather.series import Series, CategorySeries
 from leather.shapes import Bars, Columns, Dots, Line
 import leather.svg as svg
 from leather import theme
-from leather.utils import X, Y, Box, IPythonSVG
+from leather.utils import X, Y, DIMENSION_NAMES, Box, IPythonSVG
 
 
 class Chart(object):
@@ -178,7 +179,13 @@ class Chart(object):
         if not scale:
             scale = Scale.infer(self._layers, dimension, self._types[dimension])
         else:
-            scale = scale
+            for series, shape in self._layers:
+                if not scale.contains(series.min(dimension)) or not scale.contains(series.max(dimension)):
+                    d = DIMENSION_NAMES[dimension]
+                    warnings.warn('Data contains values outside %s scale domain. All data points may not be visible on the chart.' % d)
+
+                    # Only display once per axis
+                    break
 
         if not axis:
             axis = Axis()